Add VideoActivityUpdate protobuff
This proto is used for chrome to communicate updates about video playing to
powerd. This is replacing the existing message which is D-Bus int64. The new
proto has an int64 and a boolean. This CL also removes the signatures for a
related method that was never implemented. Since it was not implemented there is
no need to deprecate it first.

BUG=chrome-os-partner:9203
TEST=This is a non-functional change. This CL will be tested as part of the
     testing of the CLs that depend on it.

Change-Id: I908b2f9c2ebaccbf7a76f26edda971e9b4f2273f
Reviewed-on: https://gerrit.chromium.org/gerrit/31916
Commit-Ready: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>