Add support for new VideoActivityUpdate protobuf
This CL adds to HandleVideoActivityMethod to ability to use the new
VideoActivityUpdate protobuf that Chrome will soon be sending. Currently powerd
doesn't use all of the fields in the buffer, follow on CLs will add that
in. This is meant just to support the existing code and future CLs will add
usage of the new fields.

This CL is paired with http://codereview.chromium.org/10905026/

BUG=chrome-os-partner:9203
TEST=Test with with paired CL to confirm that there is no error messages related
     to the Video method call when video is playing.
     Test with without paired CL to confirm that there is no error messages
     related to the Video method call when video is playing.

Change-Id: I75e40ba7381c28d3e6ede7629c93a95da0e5cc19
Reviewed-on: https://gerrit.chromium.org/gerrit/32092
Tested-by: Ryan Harrison <rharrison@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>