Add KeyboardBacklightController class
This class contains the controlling logic for keyboard backlights. It uses
information from the ALS to determine when to to enable the backlight and to
what level. If there is video playing and it is fullscreen the backlight is
disabled, since we are assuming that the user is doing something like watching a
movie so does not want this backlight on.

This is the initial v0 of this controller. I just want to get all the basic
plumbing setup and all the knobs tested. Further CLs will be used to add in all
of the bells & whistles.

This CL also removes the old non-protobuf version of the HandleVideoActivity
code path, since this code is dependent on the changes in the new code path. The
new code path has been landed in Chrome and rolled into ChromeOS builds.

BUG=chrome-os-partner:9203
TEST=Build and run power_manager unittests for platorm that uses internal and
     keyboard backlight controllers.
     Build and run power_manager unittests on a platform that use the external
     backlight controller. (stumpy)
     Build and run power_manager unittests for platorm that uses internal, but
     not keyboard backlight controller. (lumpy)
     Confirm that going in and out of darkness causes the keyboard backlight to
     be turned on and off.
     Confirm that the backlight turns off when viewing youtube
     fullscreen'd.
     Test for regressions on the other two backlight combinations. (stumpy &
     lumpy)

Change-Id: I9f069a6bc45696b17fb16f52132dc8bf52ec108c
Reviewed-on: https://gerrit.chromium.org/gerrit/32971
Commit-Ready: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>