chromeos-install: Fixed parted not being found
We call parted with sudo so that we do not get an error when we run
chromeos-install without root.

I also redirect the error output of parted to /dev/null and check the return
value of expr with -z instead of comparing to 0 which was wrong.

BUG=chromium-os:34142
TEST=/usr/sbin/chromeos-install as chronos

Change-Id: I8e5332668d9350dbb7dbf871e3e54585add9d57b
Reviewed-on: https://gerrit.chromium.org/gerrit/32193
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Ready: Derek Basehore <dbasehore@chromium.org>