Fix parsing of VideoActivityUpdate protobuf
The D-Bus API uses a void *, which resolves to a char** for popping arrays. The
data structure that needs to passed in to store the array is the address of a
char* not the address of a char**, which is what the code is currently
doing. This is what led to the breakage when I switched over to using the
protobuffer path.

I was able to reproduce the crash caused by
http://codereview.chromium.org/10905026/ and this CL resolves it.

BUG=chrome-os-partner:9203
TEST=Build image with https://chromiumcodereview.appspot.com/10916123 included
     and confirm that powerd is not crashing when video is playing.

Change-Id: I8825f11d57875a5bf992091442bff998cec1969a
Reviewed-on: https://gerrit.chromium.org/gerrit/32243
Reviewed-by: Daniel Erat <derat@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>