shill: Connection: Do non-empty tests independently for DNS search
Before this change, the DNS domain search was overridden if the
DNS server paramater was non-empty.  Change this so that it is
only overridden if the derived DNS domain search list is non-empty.
This is due to reports that on DHCP renewal some servers return
a non-empty server list but do not renew the DNS domain value.

BUG=chromium-os:34260
TEST=Rerun unit tests.

Change-Id: I0f95808011e13095e467c1440dce7a13059075ad
Reviewed-on: https://gerrit.chromium.org/gerrit/32952
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>