shill: WiFiService: Don't ignore Passphrase for Configure
Do not ignore the Passphrase parameter when configuring a WiFi
service.  This allows the Manager::ConfigureService to change
the passphrase on a service which is found by GUID.

Conversely, we should not directly set the passphrase for a
WiFi service in WiFi::GetService, and instead rely on the Configure
call to handle applying the passphrase.  We can now make
WiFiService::SetPassphrase private, since there are no remaining
non-unittest consumers for this method.

BUG=chromium-os:34004
TEST=Unit-tests, small modification network_WiFiManager/095_ProfileGUID
Full WiFiManager suite.

Change-Id: Ie23c1b03dee89aea54a4a86c46734165488a885b
Reviewed-on: https://gerrit.chromium.org/gerrit/32997
Commit-Ready: Paul Stewart <pstew@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>