Add in percentage based low battery thresholds.
The EC for some devices use percentage values not time values for determining
when the device is critically low on battery. This has led to cases where hard
shutdowns have occured before the software threshold is met. This CL adds in
support for having both time and percentage based thresholds used depending on
what pref files are present.

The for the time remaining values are adjusted depending on which threshold is
being used. For the battery percentage the user visiable value is adjusted only
when using a percentage threshold.

BUG=chrome-os-partner:13899
TEST=Build and run unittests.
     Echo values into the pref files to make the threshold be very high and
     confirm that after powerd restarts the device shuts down quickly.

Change-Id: I045406bd80b82c5d7b689add2ea7c149b3acdf79
Reviewed-on: https://gerrit.chromium.org/gerrit/33138
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>