shill: Service: Emit a signal when IPConfig changes
Emit a property-changed signal whenever the Connection changes
on a Service.  This will allow listeners to detect changes to
IP address or DNS bindings.  No signal is emitted when the
Connection is reset to NULL.  The signal is emitted even if
the IPConfig path does not change, so listeners need to listen
explicitly for a PropertyChanged signal on the service, and
not simply for the value of the property changing.

BUG=chromium:134553
TEST=Manual: Do "dbus-monitor --system" and restart shill, observe:
signal sender=:1.489 -> dest=(null destination) serial=35 path=/service/1; interface=org.chromium.flimflam.Service; member=PropertyChanged
   string "IPConfig"
   variant       object path "/ipconfig/eth1_0_dhcp"

Change-Id: I5f1ab49358c53ce20bceff84c1c26453a6f8aee6
Reviewed-on: https://gerrit.chromium.org/gerrit/33342
Commit-Ready: Paul Stewart <pstew@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>