Fix chapsd upstart job start and stop conditions.
The 'ui' job has a post-stop script that depends indirectly on
chapsd.  In order for 'ui' to shut down cleanly, chapsd must remain
running until the 'stopping boot-services' event.  This fix changes
the 'chapsd' job so that a) it only depends on the public
boot-services interface, and that b) the chaps daemon stays running
long enough for a clean shutdown.

BUG=chromium-os:33354
TEST=observe no errors logged by chromeos_shutdown
TEST=run platform_BootPerfServer, and check both startup and shutdown times

Change-Id: Ib63268b547bf9ef3dc9de64149421a4fe6a5b2ff
Reviewed-on: https://gerrit.chromium.org/gerrit/33364
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Scott James Remnant <keybuk@chromium.org>
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>