Add UMA stats for suspend status
This adds a UMA stat that records the status of each suspend event.
There are three buckets currently: success, failure, and cancel. This
is all we're able to differentiate in powerd_suspend currently.

If we find that suspend failures occur frequently we could later make
this much more detailed by recovering errno from the write to
/sys/power/state, and adding a bucket per errno value.

TEST=suspend machine, check for output at chrome://histograms/SuspendStatus
BUG=chromium-os:33913
Change-Id: I214f410330fef99039d80f7359019cb3296375ec
Reviewed-on: https://gerrit.chromium.org/gerrit/33537
Reviewed-by: Darin Petkov <petkov@chromium.org>
Commit-Ready: Michael Spang <spang@chromium.org>
Tested-by: Michael Spang <spang@chromium.org>