Improve power state machine behaviour for keyboard backlight
Currently the KBC just ignores changes to power state. This CL adds in handling
them so when the device idles, powers down, etc the keyboard backlight can take
the appropriate action. These states are also sent to the screen backlight, so
the keyboard should take action at the same time as the screen backlight.

NOTE: Currently when resuming from suspend the backlight is driven to full for a
brief period then goes to teh correct value. I am confident that this is an
issue with the EC and not a software bug. I have filed crosbug.com/p/14477 about
this.

BUG=chrome-os-partner:9203
TEST=Test that the keyboard backlight dims when the device idles.
     Test that the user set brightness returns after leaving idle.
     Test that when the user turns off the backlight dimming doesn't relight it.
     Test that when the device suspends and returns the keyboard backlight turns
     off and returns to the previous state.

Change-Id: Idac302ee7ded5f15182932b9aaa12e80ca6ca59c
Reviewed-on: https://gerrit.chromium.org/gerrit/33611
Commit-Ready: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>