Refactor AmbientLightSensor to use an Observer class
Since there are potentially two controllers that might be concerned with the
response from the ALS it makes sense for it to implement the Observer
pattern. This also resolves the weird circular dependency that existed between
the controller and the sensor classes. This also adds in a get method for the
raw lux value from the ALS. This call will be used by the KBC once its full ALS
behviour is implemented.

BUG=chrome-os-partner:9203
BUG=chromium-os:34407
TEST=Build and run unittests on lumpy, stumpy and link.
     Build images for backlight and non-backlight keyboards and confirm no ALS
     regressions on the screen behaviour.

Change-Id: I422678aa96cbe9fd4f588e8940ff177e4993c3f1
Reviewed-on: https://gerrit.chromium.org/gerrit/33632
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>