Remove touch device handling around monitor reconfigure.
The disabling of touch devices on LCD power mode change
is not necessary. Still need the changes to disable the touch
devices when the lid is closed though before suspend, and when the
lid is opened, after resume.

Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=chrome-os-partner:12042,chrome-os-partner:12318
TEST=Log in as a user. Enable Show HUD for touch points.
Log back in. Suspend the system with a lid close.
Resume the system with a lid open. Ensure that the hud console in
the top left is clear of stray touch points from the lid
close/open.

Change-Id: I08f5420b8d7df0eb8dece7631fa27b8c44812f31
Reviewed-on: https://gerrit.chromium.org/gerrit/33679
Reviewed-by: Yufeng Shen <miletus@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>
Commit-Ready: Benson Leung <bleung@chromium.org>