shill: Service: Reset has_ever_connected on Unload()
This property should be removed if credentials have been reset
for this service.  While we are at it, remove a couple of bool
fields that never amounted to anything.  Add logic to test
has_ever_connected on both load and unload.  While here test
that explicitly_disconnected is cleared on Load().

BUG=chromium-os:25226
TEST=Unit tests.

Change-Id: I20580c3f78cf05db90ac1b709ac946e6cb9b6241
Reviewed-on: https://gerrit.chromium.org/gerrit/33796
Commit-Ready: Paul Stewart <pstew@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>