Autotest: Setup stumpy to power on after power failure
Added a new upstart script for test images to set up power on after
power failure for stumpy devices in the lab. It first checks that
this is indeed a lab machine and if it is also a stumpy it then sets
the correct pci bit for power on after power failure.

BUG=chromium-os:34477
TEST=Place .labmachine on my stumpy's stateful partition then ensured
the feature worked upon reboot.

Change-Id: Ib972e164b8955a9f274a683b36158d18bd7f5df7
Reviewed-on: https://gerrit.chromium.org/gerrit/33803
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>