Remove --display argument to touch-control.sh
The touch-control script no longer has a --display
flag, so remove it from the launch commands here.

BUG=chrome-os-partner:12042,chrome-os-partner:12318
TEST=Log in as a user. Enable Show HUD for touch points.
Log back in. Suspend the system with a lid close.
Resume the system with a lid open. Ensure that the hud console in
the top left is clear of stray touch points from the lid
close/open.

Change-Id: If7ce51409356db16c0a29104fc24b3cd54f39fd2
Signed-off-by: Benson Leung <bleung@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/33849
Reviewed-by: Yufeng Shen <miletus@chromium.org>