Implement ambient light behaviour in KeyboardBacklightController
Implement control of the backlight level to be a function of user input, video
state, and ambient light state. The user input is converted to be a toggle
on/off instead of direct level control. The video control remains as a
toggle. The ambient light is used to control the brightness level. There are
pref files added for the brightness steps used by the ALS and the
|target_percent_| control values.

BUG=chrome-os-partner:9203
TEST=Build and run unittests on lumpy, stumpy, and link.
     Test that Alt + BrightnessUp/Down toggles the keyboard backlight on/off.
     Test that in a pitch black room the keyboard backlight is on dimly and as
     the ambient brightness increases so does it until you are at a normal
     office level at which point it turns off.
     Test that small short changes in ambient brightness don't cause the
     backlight to jump around in its level.

Change-Id: Iafe148835b244d92a7293c5d6eafa1191225469c
Reviewed-on: https://gerrit.chromium.org/gerrit/33915
Tested-by: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Daniel Erat <derat@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>