Reland "Refactor AmbientLightSensor to use an Observer class"
The original version of this CL had a bug where in some cases powerd did not
alert the ALS that the system had entered the ACTIVE state and to start sending
ALS updates. This also meant that the state of the device was not set correctly
so exiting idle didn't do the correct things for the backlights. I have
corrected this flaw.

Since there are potentially two controllers that might be concerned with the
response from the ALS it makes sense for it to implement the Observer
pattern. This also resolves the weird circular dependency that existed between
the controller and the sensor classes. This also adds in a get method for the
raw lux value from the ALS. This call will be used by the KBC once its full ALS
behviour is implemented.

BUG=chrome-os-partner:9203
BUG=chromium-os:34407
TEST=Build and run unittests on lumpy, stumpy and link.
     Build images for backlight and non-backlight keyboards and confirm no ALS
     regressions on the screen behaviour.

Change-Id: I399a8c570ab0384388d8176f7939f23e38983dea
Reviewed-on: https://gerrit.chromium.org/gerrit/33870
Tested-by: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Daniel Erat <derat@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>