chromeos_startup: notify user of interrupted wipe
If a wipe was interrupted, the clobber-state reset file (confusingly
named /mnt/stateful_partition/factory_install_reset) would still exist.
On the next reboot, this interruption is correctly detected, but there
was no notification to the user what was going on. In this case, the
best description is that the system needs to repair itself, and this
is the message now displayed.

BUG=chrome-os-partner:13534
TEST=parrot build, manual testing

Change-Id: Ib2745c31b9049978e0e9ed352f6d2b8abaaa38ca
Signed-off-by: Kees Cook <keescook@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/34389
Reviewed-by: Jay Srinivasan <jaysri@chromium.org>