Fix the optical drive check when ejecting devices on unmount.
This CL modifies DiskManager to maintain a mapping of mount paths to
device files that should be ejected on unmount. This avoids the need to
determine if a device is an optical drive through the ID_CDROM udev
property, which is not set at the time DiskManager checks it.

BUG=chrome-os-partner:12608
TEST=Tested the following:
1. Build cros-disks for {x86,amd64,arm}-generic.
2. Run cros-disks unit tests.
3. Run platform_CrosDisksDBus and platform_CrosDisksFilesystem tests.
4. Manually verify that the DVD driver ejects a disc when:
   - Clicking the eject icon from File Manager.
   - Pressing the eject button on the DVD drive only once.

Change-Id: I3ffa7c13fd6be8e507c757d5b6afc09055b18f1b
Reviewed-on: https://gerrit.chromium.org/gerrit/34592
Commit-Ready: Ben Chan <benchan@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>