Preserve the lockbox when Powerwashing.
Before we had decided to only preserve the attestation file for
powerwashing as it was to only be used on consumer devices. This
decision has been reversed and now we will also preserve the lockbox
as well.

Now the code can preserve any file within the stateful partition and
will maintain permissions for the file and its parent directories.

BUG=chromium-os:35071
TEST=Tested the file preservation and lockbox relocation on lumpy.

Change-Id: I29e0acc72366b2eaed3ddaccc99a4af2eb9f65d9
Reviewed-on: https://gerrit.chromium.org/gerrit/34823
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Ready: Simran Basi <sbasi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>