Prevent bad time values from being sent to AddSample or Chrome
There were modications on UpdateAveragedTime to support this. If a raw time
value is negative, then we skip the entire update and reschedule in 5 secs. If
the time remaining value minus the threshold is negative then it is clamped to 0
and a warning message is put in the log. If the raw time value of interest is
above a day worth of seconds then we skip the update and reschedule. When we
skip an update we also mark the status as calculating and set the averaged times
to 0, so that the UX knows the update is bad if it queries for the status before
the next update.

BUG=chrome-os-partner:14758
TEST=Built and ran unittests.
     Built image and confirmed that battery times are sane around events like
     pulling the power plug and resuming.
     Used temp logging to confirm that the rolling averages are not getting bad
     values.

Change-Id: Ifac94a46099776533f94c5cb64a85e93918cf96d
Reviewed-on: https://gerrit.chromium.org/gerrit/35035
Commit-Ready: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>