Add a stop condition to the openssh-server job
The lack of stop condition means that sshd stays running well
past the time that most processes are terminated.  That's not
a good idea.

The lack of stop is also suspected to be causing anomalies running
boot time tests in the test lab, although that isn't proven (yet).

BUG=chromium-os:34877
TEST=platform_BootPerfServer

Change-Id: Ia047118ad6afb1041b1814725702bd8721fa0f1a
Reviewed-on: https://gerrit.chromium.org/gerrit/34966
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Scott James Remnant <keybuk@chromium.org>
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>