Add additional logging for KeyboardBacklightController
This adds in a series of logging statements in KBC so that regular logs have
more detail on what is occuring. This also refactors the ALS response history
code in InternalBacklightController, so that is now in AmbientLightSensor and
accessiable to all Observers of that class instead of having them reimplement
the functionality.

BUG=chrome-os-partner:15332
TEST=Looked in the log and made sure I could figure out what was going on from
     the logs and that the messages were not spammy.

Change-Id: Ifeb0d3af25e7a111ae061e6fc1677309568c6969
Reviewed-on: https://gerrit.chromium.org/gerrit/35847
Commit-Ready: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>