Add metric for when bad battery state values are read from sysfs
This metric is used to count the number of times that we read values from sysfs
for updating power information and the number of times that read is good/bad. On
some devices we have had issues with bad reads after power events like
starting/resuming. These have been mitigated in software and there are fixes
being worked on for the specific platforms in question. This metric is being
added so that we can track with new devices being brought up if they are having
similar issues.

Additional unittests are added to cover all use cases of this metric. This code
also refactors another metric that has a similar implementation so that they
have a common unified style. There is also a slight refactoring of the
UpdateAveragedTimes method since the part I was bringing under test didn't fit
stylistically.

BUG=chromium-os:35550
TEST=Run unittests.
     Build image and confirm that after atleast 30 seconds chrome://histograms/
     has instances of Power.BatteryInfoSample.

Change-Id: I190f1faa5966c06a76c4d73b2f073e7147fc4811
Reviewed-on: https://gerrit.chromium.org/gerrit/36434
Tested-by: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Daniel Erat <derat@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>